datapath/fake: Move MTU to fake/types package
See previous commit for rationale.
