+++ pkg/datapath/fake/types/mtu.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package fake
+package types
 
 import "github.com/cilium/cilium/pkg/mtu"
